organizes directory
