--- src/components/Dashboard/components/Data.js
@@ -1,80 +1,0 @@
-export const data = {
-  Breakfast: [
-    {
-      food: "Oatmeal",
-      serving: 1,
-      fat: 5.3,
-      protein: 10.7,
-      carbs: 46.7,
-      calories: 307,
-    },
-    {
-      food: "Raspberries",
-      serving: 1,
-      fat: 2.0,
-      protein: 2.9,
-      carbs: 20.5,
-      calories: 140,
-    },
-    {
-      food: "Black Coffee",
-      serving: 1,
-      fat: 0.4,
-      protein: 0.3,
-      carbs: 20.5,
-      calories: 21,
-    },
-  ],
-  Lunch: [
-    {
-      food: "Sweet Potato",
-      serving: 1,
-      fat: 0.2,
-      protein: 3.0,
-      carbs: 26.0,
-      calories: 135,
-    },
-    {
-      food: "Tossed Salad",
-      serving: 2,
-      fat: 0.2,
-      protein: 1.0,
-      carbs: 2.0,
-      calories: 16,
-    },
-    {
-      food: "Avocado",
-      serving: 2,
-      fat: 4.4,
-      protein: 0.5,
-      carbs: 0.6,
-      calories: 48,
-    },
-  ],
-  Dinner: [
-    {
-      food: "Grilled Salmon",
-      serving: 3,
-      fat: 6.0,
-      protein: 22.0,
-      carbs: 0,
-      calories: 150,
-    },
-    {
-      food: " Steamed Broccoli",
-      serving: 2,
-      fat: 1.3,
-      protein: 7.4,
-      carbs: 12.1,
-      calories: 109,
-    },
-    {
-      food: "Brown Rice",
-      serving: 1,
-      fat: 2.0,
-      protein: 5.5,
-      carbs: 48.4,
-      calories: 248,
-    },
-  ],
-};
