+++ src/components/Dashboard/components/FoodJournal/Favorites/Favorites.js
