+++ src/components/Dashboard/components/FoodJournal/MealEntries.js
@@ -1,7 +1,7 @@
 import React from "react";
-import { ReactComponent as FullHeart } from "../../../../../assets/FullHeart.svg";
-import { ReactComponent as Heart } from "../../../../../assets/HeartOutline.svg";
-import { ReactComponent as TrashCan } from "../../../../../assets/Trash-can.svg";
+import { ReactComponent as FullHeart } from "../../../../assets/FullHeart.svg";
+import { ReactComponent as Heart } from "../../../../assets/HeartOutline.svg";
+import { ReactComponent as TrashCan } from "../../../../assets/Trash-can.svg";
 
 const MealEntries = (props) => {
   return (
@@ -25,7 +25,10 @@ const MealEntries = (props) => {
                     <Heart onClick={props.toggleFavorite} />
                   )}
                 </span>
-                <span className="font-weight-bold text-capitalize">{entry.food_name},</span> {entry.measurement_name}
+                <span className="font-weight-bold text-capitalize">
+                  {entry.food_name},
+                </span>{" "}
+                {entry.measurement_name}
               </td>
               <td>{entry.quantity}</td>
               <td>{entry.fat_g} g</td>
