+++ src/components/Dashboard/components/FoodSearch/SearchResults.js
@@ -1,8 +1,8 @@
 import React, { useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { ReactComponent as Placeholder } from "../../../assets/undraw_breakfast.svg";
-import { getNutrients, searchFood } from "../../../state/slices/EdamamSlice";
-import NutritionInfo from "./Nutrition/NutritionInfo";
+import { ReactComponent as Placeholder } from "../../../../assets/undraw_breakfast.svg";
+import { getNutrients } from "../../../../state/slices/EdamamSlice";
+import NutritionInfo from "../Nutrition/NutritionInfo";
 import SearchResultsList from "./SearchResultsList";
 
 const SearchResults = () => {
