+++ src/components/Dashboard/components/FoodSearch/SearchResultsList.js
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import { useSelector, useDispatch } from "react-redux";
+import { useDispatch, useSelector } from "react-redux";
 import Pagination from "./Pagination";
 
 const SearchResultsList = (props) => {
