--- src/components/Dashboard/components/ProgressBar.js
@@ -1,14 +1,0 @@
-import React from "react";
-
-var progressBar = require("progressbar.js");
-
-let calories = new progressBar.Circle(calories, {
-  color: "#aaa",
-  strokWidth,
-});
-
-let protein = new progressBar.Circle(protein);
-
-let fat = new progressBar.Circle(fat);
-
-let carbs = new progressBar.Circle(carbs);
